lang-chain-mcp-adapters/pyproject.toml
[project]
name = "lang-chain-mcp-adapters"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.16.1",
    "isort>=6.0.1",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.4",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-openai>=0.3.17",
    "langgraph>=0.4.5",
    "langsmith>=0.1.0",
    "mcp>=1.9.0",
    "python-dotenv>=1.1.0",
]
